verify test case for python works now
